refactor(upscaling): move OpenComposite detection to VRDetection
Extract the opencomposite.ini probing/parsing out of Upscaling.cpp into
VRDetection::DetectOpenCompositeUpscaling, where it sits alongside the
existing OpenComposite runtime detection. Upscaling keeps only the
force-to-None policy and the lifecycle/UI gating, dropping ~230 lines
and its parallel blocker struct in favor of the shared VRDetection type.

Consolidate the loaded openvr_api.dll path lookup into one shared helper
(GatherDLLInfo previously had its own copy), and search
opencomposite_ext.ini next to the DLL too, not just in the CWD.
