+++ src/Features/Upscaling.cpp
@@ -16,14 +16,10 @@
 #include "Utils/UI.h"
 #include <Windows.h>
 #include <algorithm>
-#include <cctype>
 #include <cfloat>
 #include <cmath>
-#include <cwctype>
 #include <directx/d3dx12.h>
-#include <filesystem>
 #include <format>
-#include <string_view>
 
 NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE_WITH_DEFAULT(
 	Upscaling::Settings,
@@ -186,232 +182,6 @@ HRESULT WINAPI hk_D3D11CreateDeviceAndSwapChainUpscaling(
 	}
 
 	return ret;
-}
-
-namespace
-{
-	// OpenComposite (OpenXR-over-OpenVR shim) can run its own DLSS/FSR/DLAA
-	// upscaling. If our upscaler also runs, the two stack — double jitter,
-	// double upscale, conflicting dynamic-resolution sizing — which wrecks the
-	// VR image. We detect OpenComposite's upscaling from its config and stand
-	// down so it owns the upscale. VR-only; SteamVR users are unaffected.
-	struct OpenCompositeSettingValue
-	{
-		bool value = false;
-		std::string configPath;
-	};
-
-	struct OpenCompositeUpscalingSettings
-	{
-		OpenCompositeSettingValue dlssEnabled;
-		OpenCompositeSettingValue fsrEnabled;
-		OpenCompositeSettingValue dlaaEnabled;
-		OpenCompositeSettingValue fsrNativeAA;
-		OpenCompositeSettingValue fsr3PostAAEnabled;
-	};
-
-	struct DetectedOpenCompositeUpscalingBlocker
-	{
-		bool active = false;
-		std::string settingName;
-		std::string configPath;
-	};
-
-	std::string_view TrimAsciiWhitespace(std::string_view value)
-	{
-		while (!value.empty() && std::isspace(static_cast<unsigned char>(value.front())))
-			value.remove_prefix(1);
-		while (!value.empty() && std::isspace(static_cast<unsigned char>(value.back())))
-			value.remove_suffix(1);
-		return value;
-	}
-
-	std::string ToLowerAscii(std::string_view value)
-	{
-		std::string result(value);
-		std::ranges::transform(result, result.begin(), [](unsigned char c) {
-			return static_cast<char>(std::tolower(c));
-		});
-		return result;
-	}
-
-	bool TryParseOpenCompositeBool(std::string value, bool& outValue)
-	{
-		value = ToLowerAscii(TrimAsciiWhitespace(value));
-		if (value == "true" || value == "on" || value == "enabled" || value == "1") {
-			outValue = true;
-			return true;
-		}
-		if (value == "false" || value == "off" || value == "disabled" || value == "0") {
-			outValue = false;
-			return true;
-		}
-		return false;
-	}
-
-	void AddUniquePath(std::vector<std::filesystem::path>& paths, const std::filesystem::path& path)
-	{
-		if (path.empty())
-			return;
-
-		auto normalized = path.lexically_normal().wstring();
-		std::ranges::transform(normalized, normalized.begin(), [](wchar_t c) {
-			return static_cast<wchar_t>(std::towlower(c));
-		});
-
-		const bool alreadyAdded = std::ranges::any_of(paths, [&](const std::filesystem::path& existing) {
-			auto existingNormalized = existing.lexically_normal().wstring();
-			std::ranges::transform(existingNormalized, existingNormalized.begin(), [](wchar_t c) {
-				return static_cast<wchar_t>(std::towlower(c));
-			});
-			return existingNormalized == normalized;
-		});
-		if (!alreadyAdded)
-			paths.push_back(path);
-	}
-
-	std::filesystem::path GetCurrentDirectoryPath()
-	{
-		std::wstring buffer(MAX_PATH, L'\0');
-		const DWORD length = GetCurrentDirectoryW(static_cast<DWORD>(buffer.size()), buffer.data());
-		if (length == 0)
-			return {};
-
-		if (length >= buffer.size()) {
-			buffer.resize(length + 1);
-			const DWORD retryLength = GetCurrentDirectoryW(static_cast<DWORD>(buffer.size()), buffer.data());
-			if (retryLength == 0 || retryLength >= buffer.size())
-				return {};
-			buffer.resize(retryLength);
-		} else {
-			buffer.resize(length);
-		}
-
-		return std::filesystem::path(buffer);
-	}
-
-	// Resolves the directory of the already-loaded openvr_api.dll. Self-contained
-	// rather than reusing VRDetection::GatherDLLInfo, which also parses version
-	// info and enumerates the file — more work than the directory lookup needs.
-	std::filesystem::path GetLoadedOpenVRDirectory()
-	{
-		HMODULE openVRModule = GetModuleHandleW(L"openvr_api.dll");
-		if (!openVRModule)
-			return {};
-
-		std::wstring buffer(MAX_PATH, L'\0');
-		const DWORD length = GetModuleFileNameW(openVRModule, buffer.data(), static_cast<DWORD>(buffer.size()));
-		if (length == 0 || length >= buffer.size())
-			return {};
-
-		buffer.resize(length);
-		return std::filesystem::path(buffer).parent_path();
-	}
-
-	std::vector<std::filesystem::path> GetOpenCompositeConfigCandidates()
-	{
-		std::vector<std::filesystem::path> candidates;
-
-		const auto loadedOpenVRDirectory = GetLoadedOpenVRDirectory();
-		if (!loadedOpenVRDirectory.empty())
-			AddUniquePath(candidates, loadedOpenVRDirectory / L"opencomposite.ini");
-
-		const auto currentDirectory = GetCurrentDirectoryPath();
-		if (!currentDirectory.empty()) {
-			AddUniquePath(candidates, currentDirectory / L"opencomposite.ini");
-			AddUniquePath(candidates, currentDirectory / L"opencomposite_ext.ini");
-		}
-
-		return candidates;
-	}
-
-	// OpenComposite writes upscaling keys at top level or under a section
-	// depending on build, so probe the unnamed section first then every named
-	// one.
-	bool TryReadIniBoolSetting(const CSimpleIniA& ini, const char* key, bool& outValue)
-	{
-		auto tryReadSection = [&](const char* section) {
-			const char* rawValue = ini.GetValue(section, key, nullptr);
-			return rawValue && TryParseOpenCompositeBool(rawValue, outValue);
-		};
-
-		if (tryReadSection(""))
-			return true;
-
-		CSimpleIniA::TNamesDepend sections;
-		ini.GetAllSections(sections);
-		for (const auto& section : sections) {
-			if (section.pItem && tryReadSection(section.pItem))
-				return true;
-		}
-
-		return false;
-	}
-
-	void UpdateOpenCompositeSettingValue(OpenCompositeSettingValue& setting, const CSimpleIniA& ini, const char* key, const std::filesystem::path& path)
-	{
-		bool parsedValue = false;
-		if (!TryReadIniBoolSetting(ini, key, parsedValue))
-			return;
-
-		setting.value = parsedValue;
-		setting.configPath = path.string();
-	}
-
-	OpenCompositeUpscalingSettings ReadOpenCompositeUpscalingSettings()
-	{
-		OpenCompositeUpscalingSettings settings;
-
-		std::error_code ec;
-		for (const auto& path : GetOpenCompositeConfigCandidates()) {
-			if (!std::filesystem::exists(path, ec))
-				continue;
-			ec.clear();
-
-			CSimpleIniA ini;
-			ini.SetUnicode();
-			const SI_Error rc = ini.LoadFile(path.c_str());
-			if (rc < 0) {
-				logger::warn("[Upscaling] Failed to read OpenComposite config '{}': {}", path.string(), static_cast<int>(rc));
-				continue;
-			}
-
-			UpdateOpenCompositeSettingValue(settings.dlssEnabled, ini, "dlssEnabled", path);
-			UpdateOpenCompositeSettingValue(settings.fsrEnabled, ini, "fsrEnabled", path);
-			UpdateOpenCompositeSettingValue(settings.dlaaEnabled, ini, "dlaaEnabled", path);
-			UpdateOpenCompositeSettingValue(settings.fsrNativeAA, ini, "fsrNativeAA", path);
-			UpdateOpenCompositeSettingValue(settings.fsr3PostAAEnabled, ini, "fsr3PostAAEnabled", path);
-		}
-
-		return settings;
-	}
-
-	DetectedOpenCompositeUpscalingBlocker FindOpenCompositeUpscalingBlocker()
-	{
-		DetectedOpenCompositeUpscalingBlocker blocker;
-		if (!globals::game::isVR)
-			return blocker;
-
-		const auto settings = ReadOpenCompositeUpscalingSettings();
-		auto setBlocker = [&](const char* settingName, const OpenCompositeSettingValue& setting) {
-			blocker.active = true;
-			blocker.settingName = settingName;
-			blocker.configPath = setting.configPath;
-		};
-
-		if (settings.dlaaEnabled.value)
-			setBlocker("dlaaEnabled", settings.dlaaEnabled);
-		else if (settings.fsrNativeAA.value)
-			setBlocker("fsrNativeAA", settings.fsrNativeAA);
-		else if (settings.fsr3PostAAEnabled.value)
-			setBlocker("fsr3PostAAEnabled", settings.fsr3PostAAEnabled);
-		else if (settings.dlssEnabled.value)
-			setBlocker("dlssEnabled", settings.dlssEnabled);
-		else if (settings.fsrEnabled.value)
-			setBlocker("fsrEnabled", settings.fsrEnabled);
-
-		return blocker;
-	}
 }
 
 void Upscaling::DrawSettings()
@@ -851,17 +621,17 @@ void Upscaling::DrawSettings()
 	}
 }
 
-const Upscaling::OpenCompositeUpscalingBlocker& Upscaling::GetOpenCompositeUpscalingBlocker(bool a_forceRefresh) const
+const VRDetection::OpenCompositeUpscalingState& Upscaling::GetOpenCompositeUpscalingBlocker(bool a_forceRefresh) const
 {
 	// Cached after the first probe; lifecycle entry points pass forceRefresh so
 	// per-frame callers (GetUpscaleMethod, DrawSettings) only read the bool.
 	if (!a_forceRefresh && openCompositeUpscalingBlockerCacheValid)
 		return openCompositeUpscalingBlocker;
 
-	const auto detected = FindOpenCompositeUpscalingBlocker();
-	openCompositeUpscalingBlocker.active = detected.active;
-	openCompositeUpscalingBlocker.settingName = detected.settingName;
-	openCompositeUpscalingBlocker.configPath = detected.configPath;
+	// Only OpenComposite VR users can hit this conflict — skip the probe otherwise.
+	openCompositeUpscalingBlocker = globals::game::isVR ?
+	                                    VRDetection::DetectOpenCompositeUpscaling() :
+	                                    VRDetection::OpenCompositeUpscalingState{};
 	openCompositeUpscalingBlockerCacheValid = true;
 
 	return openCompositeUpscalingBlocker;
