+++ src/Features/VR/OpenVRDetection.h
@@ -32,6 +32,16 @@ namespace VRDetection
 		bool probingSucceeded = false;
 	};
 
+	// Whether OpenComposite is running its own DLSS/FSR/DLAA upscaling, read
+	// from its config. Lets a feature (e.g. Upscaling) stand down so the two
+	// don't stack. settingName/configPath identify the trigger for UI/logs.
+	struct OpenCompositeUpscalingState
+	{
+		bool active = false;
+		std::string settingName;
+		std::string configPath;
+	};
+
 	// Runtime interface probing via VR_IsInterfaceVersionValid
 	bool ProbeRuntimeInterfaces(OpenVRDetectionResult& result);
 
@@ -43,6 +53,12 @@ namespace VRDetection
 
 	// Full detection via interface probing
 	OpenVRDetectionResult Detect();
+
+	// Probe opencomposite.ini / opencomposite_ext.ini (next to the loaded
+	// openvr_api.dll and in the CWD) for an enabled DLSS/FSR/DLAA upscaler.
+	// Returns an inactive state when no config or no upscaler is found. Does
+	// not itself check whether VR is active — callers gate on that.
+	OpenCompositeUpscalingState DetectOpenCompositeUpscaling();
 
 	const char* RuntimeTypeToString(RuntimeType type);
 }
